- Removed __rewrite (I Think):
The PATH_TRANSLATED variable is only supposed to exist if rewrites have been done in some way shape or form.  Or so is my understanding.  The index "front controller" has been modified to adjust variables based on this instead of the old __rewrite param.  This is not tested but should work.