Moved config into further sub directory, added new config options such as execution mode which will determine minification type.  Made config path and config name separate.  Moved logic for / and translated paths to routing.  CLeaned up init.php by placing all routing related calls in the index.php file and cleaned up the index.php.  Removed INCLUDES_ROOT.  APPLICATION_ROOT is the only required definition and it is detected based on the location of the console or the includes folder, period.  The console assumes it's in the root, while the index script is cleaned up to loop looking for the includes directory.  Directories which are assumed in the application root are includes and config only.  The remainder should all be configurable relative to the application root (including backwards).  It is completely possible to move everything in user back one directory, rename the directory as "app" and modify the roots and autoload paths of other files.  Minimally it would take a few minutes to have this structure if people so desired.
You can now simply copy "default" to another folder, change the name, and modify init.php to load that.

Additional mime types have been added to the controller, and the SSL logic was moved to __init() as it can be handled pre-instantiation.  It is not yet determined how the final view rendering will occur.  The ideal is to have controllers always return whatever is to be presented to the user which would mean index.php would be responsible for determining how to actually output that.